backend/requirements.txt
asgiref==3.7.2
boto3==1.34.36
botocore==1.34.36
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
jmespath==1.0.1
gunicorn==20.1.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2024.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
python-dotenv==1.0.0
djangorestframework-simplejwt